Refactor routes formation and remove from sidebar.js and add to own file, routes.js
